Can send data back and forth via button press more than once now.
Was an issue with can_read flag. Used 2 in the heat of the moment, but could probably do with one. Client reads then writes (needs to be able to read right away so can_read == true). Server does the opposite and should not read right away or itll get stuck (can_read == false)